// src/models/UtPcaClientId.ts
export class UtPcaClientId {
	clientId?: number;
	adminId?: number;
	clientName?: string;
	clientDir?: string;
	clientDbName?: string;
	clientClntGrpSrc?: string;
	clientClntGrpUtlzLnk?: string;
	clientClntGrpUtlzLnkDesc?: string;
	hpclntBidLkup?: string;
	clientEffecDt?: string;
	clientTermDt?: string;
	clientBenfTypDef?: string;
	clientPhone?: string;
	pmoIdqualifier?: string;
	pmoIdCode?: string;
	pmoIdName?: string;
	groupingManufPlanInfo?: number;
	logoImageSrc?: string;
	logoImageAlt?: string;
	defaultContProd?: number;
	defaultCont?: number;
	defaultRxDotShift?: string;
	utilDataFilePath?: string;
	clientAbbreviation?: string;
	clientType?: string;
	defaultPosition?: number;
	cycleFormatFile?: string;
	claimAuditTableFile?: string;
	disbursementBeginDate?: string;
	activeClient?: boolean;
	clientMemberLinkKey?: string;
	exagoReportDirectory?: string;
	proOptFormularyId?: Array<ProOptFormularyId>;
	utPcaHpbnftId?: Array<UtPcaHpbnftId>;
	utPcaHpbnftTypeId?: Array<UtPcaHpbnftTypeId>;
	utPcaHpclntGrpId?: Array<UtPcaHpclntGrpId>;
	utPcaHpcycleId?: Array<UtPcaHpcycleId>;
	utPcaHpdsgnId?: Array<UtPcaHpdsgnId>;
	utPcaHpfrmlyId?: Array<UtPcaHpfrmlyId>;
	utPcaHpfrmtFileId?: Array<UtPcaHpfrmtFileId>;
	utPcaHptblColMapId?: Array<UtPcaHptblColMapId>;
	utPcaPharmaId?: Array<UtPcaPharmaId>;
	utPcaResultClientId?: Array<UtPcaResultClientId>;
	utPcaRunId?: Array<UtPcaRunId>;
	constructor(obj: any = {}) {
		this.clientId = obj.clientId;
		this.adminId = obj.adminId;
		this.clientName = obj.clientName;
		this.clientDir = obj.clientDir;
		this.clientDbName = obj.clientDbName;
		this.clientClntGrpSrc = obj.clientClntGrpSrc;
		this.clientClntGrpUtlzLnk = obj.clientClntGrpUtlzLnk;
		this.clientClntGrpUtlzLnkDesc = obj.clientClntGrpUtlzLnkDesc;
		this.hpclntBidLkup = obj.hpclntBidLkup;
		this.clientEffecDt = obj.clientEffecDt;
		this.clientTermDt = obj.clientTermDt;
		this.clientBenfTypDef = obj.clientBenfTypDef;
		this.clientPhone = obj.clientPhone;
		this.pmoIdqualifier = obj.pmoIdqualifier;
		this.pmoIdCode = obj.pmoIdCode;
		this.pmoIdName = obj.pmoIdName;
		this.groupingManufPlanInfo = obj.groupingManufPlanInfo;
		this.logoImageSrc = obj.logoImageSrc;
		this.logoImageAlt = obj.logoImageAlt;
		this.defaultContProd = obj.defaultContProd;
		this.defaultCont = obj.defaultCont;
		this.defaultRxDotShift = obj.defaultRxDotShift;
		this.utilDataFilePath = obj.utilDataFilePath;
		this.clientAbbreviation = obj.clientAbbreviation;
		this.clientType = obj.clientType;
		this.defaultPosition = obj.defaultPosition;
		this.cycleFormatFile = obj.cycleFormatFile;
		this.claimAuditTableFile = obj.claimAuditTableFile;
		this.disbursementBeginDate = obj.disbursementBeginDate;
		this.activeClient = obj.activeClient;
		this.clientMemberLinkKey = obj.clientMemberLinkKey;
		this.exagoReportDirectory = obj.exagoReportDirectory;
		this.proOptFormularyId = obj.proOptFormularyId.map(o => new ProOptFormularyId(o));
		this.utPcaHpbnftId = obj.utPcaHpbnftId.map(o => new UtPcaHpbnftId(o));
		this.utPcaHpbnftTypeId = obj.utPcaHpbnftTypeId.map(o => new UtPcaHpbnftTypeId(o));
		this.utPcaHpclntGrpId = obj.utPcaHpclntGrpId.map(o => new UtPcaHpclntGrpId(o));
		this.utPcaHpcycleId = obj.utPcaHpcycleId.map(o => new UtPcaHpcycleId(o));
		this.utPcaHpdsgnId = obj.utPcaHpdsgnId.map(o => new UtPcaHpdsgnId(o));
		this.utPcaHpfrmlyId = obj.utPcaHpfrmlyId.map(o => new UtPcaHpfrmlyId(o));
		this.utPcaHpfrmtFileId = obj.utPcaHpfrmtFileId.map(o => new UtPcaHpfrmtFileId(o));
		this.utPcaHptblColMapId = obj.utPcaHptblColMapId.map(o => new UtPcaHptblColMapId(o));
		this.utPcaPharmaId = obj.utPcaPharmaId.map(o => new UtPcaPharmaId(o));
		this.utPcaResultClientId = obj.utPcaResultClientId.map(o => new UtPcaResultClientId(o));
		this.utPcaRunId = obj.utPcaRunId.map(o => new UtPcaRunId(o));
	}
}