// src/models/UtPcaContId.ts
import { UtPcaContStatusId } from './UtPcaContStatusId';
import { UtPcaPharmaId } from './UtPcaPharmaId';
export class UtPcaContId {
	contId?: number;
	pharmaId?: number;
	deletedFlag?: string;
	contNumber?: string;
	contName?: string;
	contEffecDt?: string;
	contTermDt?: string;
	contStatusId?: number;
	contNote?: string;
	contLangId?: string;
	dayPay?: number;
	daySubmit?: number;
	mfgDataGrp?: number;
	bundleId?: number;
	analystId?: number;
	inEligRuleId?: number;
	pharmaNoticeContactId?: string;
	pharmaSubmitContactId?: string;
	contractTypeId?: string;
	daysBefCont?: number;
	daysAftCont?: number;
	daysBefRun?: number;
	daysAftRun?: number;
	contCreateDate?: string;
	contUpdateDate?: string;
	frmlyDataGrp?: number;
	crossDataGrp?: number;
	lastChangeUser?: string;
	contStatus?: UtPcaContStatusId;
	pharma?: UtPcaPharmaId;
	constructor(obj: any = {}) {
		this.contId = obj.contId;
		this.pharmaId = obj.pharmaId;
		this.deletedFlag = obj.deletedFlag;
		this.contNumber = obj.contNumber;
		this.contName = obj.contName;
		this.contEffecDt = obj.contEffecDt;
		this.contTermDt = obj.contTermDt;
		this.contStatusId = obj.contStatusId;
		this.contNote = obj.contNote;
		this.contLangId = obj.contLangId;
		this.dayPay = obj.dayPay;
		this.daySubmit = obj.daySubmit;
		this.mfgDataGrp = obj.mfgDataGrp;
		this.bundleId = obj.bundleId;
		this.analystId = obj.analystId;
		this.inEligRuleId = obj.inEligRuleId;
		this.pharmaNoticeContactId = obj.pharmaNoticeContactId;
		this.pharmaSubmitContactId = obj.pharmaSubmitContactId;
		this.contractTypeId = obj.contractTypeId;
		this.daysBefCont = obj.daysBefCont;
		this.daysAftCont = obj.daysAftCont;
		this.daysBefRun = obj.daysBefRun;
		this.daysAftRun = obj.daysAftRun;
		this.contCreateDate = obj.contCreateDate;
		this.contUpdateDate = obj.contUpdateDate;
		this.frmlyDataGrp = obj.frmlyDataGrp;
		this.crossDataGrp = obj.crossDataGrp;
		this.lastChangeUser = obj.lastChangeUser;
		this.contStatus = new UtPcaContStatusId(obj.contStatus);
		this.pharma = new UtPcaPharmaId(obj.pharma);
	}
}